chore: clean up telegram reporting message and data models
